+++ include/render/primitives/ConvexOperation.h
@@ -9,6 +9,8 @@ namespace render {
                                        render::State& state) const override;
     PrimitivePacketHit4 intersectPacketHits(const Ray4& rays,
                                             const PrimitivePacketState4& states) const override;
+    PrimitivePacketHit8 intersectPacketHits(const Ray8& rays,
+                                            const PrimitivePacketState8& states) const override;
     virtual bool intersects(const Rayd& ray, render::State& state) const override;
 
     virtual Vector3d farthestPoint(const Vector3d& direction) const override = 0;